Move Annotation from Playground to the official part
